% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smEn.R
\name{smEn}
\alias{smEn}
\title{Simulador de Efeitos de Ambiente não identificável}
\usage{
smEn(nind, m, dp)
}
\arguments{
\item{nind}{número de indivíduos da população}

\item{m}{média da distribuição normal}

\item{dp}{desvio-padrão da distribuição normal}
}
\value{
um vetor com os efeito de ambiente simulados
}
\description{
Simulador de Efeitos de Ambiente não identificável
}
\examples{
smEn(nind=500,m=0,dp=25)
}
